+++ src/OhadSoft.AzureLetsEncrypt.Renewal/Management/ParamValidator.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 
 namespace OhadSoft.AzureLetsEncrypt.Renewal.Management
@@ -25,29 +24,6 @@ namespace OhadSoft.AzureLetsEncrypt.Renewal.Management
             return str == null || !str.All(Char.IsWhiteSpace)
                 ? str
                 : throw new ArgumentException("String must be either null or non-whitespace", name);
-        }
-
-        public static IReadOnlyList<string> VerifyHosts(IReadOnlyList<string> hosts, bool enforceWildcard, string name)
-        {
-            if (hosts == null || hosts.Count == 0 || hosts.Any(h => Uri.CheckHostName(h?.Replace('*', 'x')) == UriHostNameType.Unknown))
-            {
-                throw new ArgumentException("Host collection must be non-null, contain at least one element, and contain valid host names", name);
-            }
-
-            if (enforceWildcard && hosts.Any(h => !h.StartsWith("*.", StringComparison.OrdinalIgnoreCase)))
-            {
-                throw new ArgumentException("Only wildcard host names are supported for the DNS challenge (must begin with '*.'", name);
-            }
-
-            return hosts;
-        }
-
-        public static string VerifyEmail(string email, string name)
-        {
-            return !String.IsNullOrWhiteSpace(email) && email.Contains("@") && email.Length >= 3 && email.Length <= 254
-                   && !email.StartsWith("@", StringComparison.OrdinalIgnoreCase) && !email.EndsWith("@", StringComparison.OrdinalIgnoreCase)
-                ? email
-                : throw new ArgumentException("E-mail address must not be null and must be valid", name);
         }
 
         public static string VerifyString(string str, string name)
